requirements.txt
pygame
ffmpeg-python
youtube-dl
pytube